snapcraft: Add basic snap packaging
This adds a basic snapcraft.yaml which generates a simple package that
can be run on any systems with snapd.

To avoid having to deal with all the confinement bits, this relies on
the unbound TCP remote socket instead of using the unix socket or shm API.

It binds on all IPv4/IPv6 addresses on the default port.

Signed-off-by: Stéphane Graber <stgraber@ubuntu.com>